requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
datadog==0.34.1
decorator==4.4.2
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
influxdb==5.3.0
jinja2==2.11.2
markupsafe==1.1.1
msgpack==0.6.1
oauth2client==4.1.3
oauthlib==3.1.0
paho-mqtt==1.5.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybluez==0.22
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.10
